refactor(library/vue): moves "/statefulAttempt.ts" to "/composables/useStatefulAttemptThatEventually.ts"
